feat(notification): moving methods to notification app
